Move pattern constructor to models directory
