+++ src/core/models/pattern-constructor.mjs
